backend/requirements.txt
Django==5.1.7
psycopg2-binary
python-dotenv
requests
djangorestframework